% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alias.R
\name{alias_sample}
\alias{alias_sample}
\title{Sample from an alias table in O(1) time}
\usage{
alias_sample(alias_obj)
}
\arguments{
\item{alias_obj}{A list produced by \code{build_alias()},
containing elements \code{prob} and \code{alias}.}
}
\value{
An integer index corresponding to the sampled outcome.
}
\description{
Draws a single sample from a discrete distribution
represented by an alias table.
}
\details{
Each draw uses one uniform integer sample and one uniform
random number, achieving constant-time sampling.
}
\examples{
#p <- c(0.1, 0.2, 0.7)
#alias <- build_alias(p)
#alias_sample(alias)

}
