% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bipartite_to_rig.R
\name{bipartite_to_rig}
\alias{bipartite_to_rig}
\title{Bipartite to RIG}
\usage{
bipartite_to_rig(bipartite_matrix)
}
\arguments{
\item{bipartite_matrix}{bipartite_matrix}
}
\value{
Random Intersection Graph \code{rig}.
}
\description{
Bipartite to RIG
}
\examples{
#---------------------------------
# bipartite_to_rig(bipartite_matrix)
#---------------------------------
}
\author{
Victor Bernal \email{victor.arturo.bernal@gmail.com}
}
