% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alias.R
\name{build_alias}
\alias{build_alias}
\title{Build an alias table for O(1) discrete sampling}
\usage{
build_alias(p)
}
\arguments{
\item{p}{Numeric vector of probabilities (does not need to sum to 1;
it will be normalized internally).}
}
\value{
A list with two elements:
\describe{
\item{prob}{Numeric vector of probabilities for each outcome}
\item{alias}{Integer vector of alias indices}
}
}
\description{
Constructs the alias method data structure from a probability vector.
This allows sampling from a discrete distribution in constant time.
}
\details{
The alias method preprocesses a discrete distribution in O(n) time
and allows subsequent sampling in O(1) time.
}
\examples{
#p <- c(0.1, 0.2, 0.7)
#alias <- build_alias(p)

}
