% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rates.R
\name{compute_rates}
\alias{compute_rates}
\title{Rates}
\usage{
compute_rates(state, params)
}
\arguments{
\item{state}{Current simulation state.}

\item{params}{Simulation parameters.}
}
\value{
Event rates.
}
\description{
Calculate event rates for the current simulation state.
}
