% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debug_illegal_schelling.R
\docType{package}
\name{debug_illegal_schelling}
\alias{debug_illegal_schelling}
\title{Detect illegal Schelling moves in a bipartite matrix}
\usage{
debug_illegal_schelling(old_bipartite, new_bipartite, t = NULL)
}
\arguments{
\item{old_bipartite}{old_bipartite}

\item{new_bipartite}{new_bipartite}

\item{t}{time}
}
\examples{
#---------------------------------
# debug_illegal_transition
#---------------------------------
}
\references{
\url{}
}
\seealso{
\code{\link{brocolors}}
}
\author{
Victor Bernal \email{victor.arturo.bernal@gmail.com}
}
\keyword{hplot}
