% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_helpers.R
\name{get_dark_state_flags}
\alias{get_dark_state_flags}
\title{Dark-State Flags}
\usage{
get_dark_state_flags(num_opinions)
}
\arguments{
\item{num_opinions}{Number of opinion states.}
}
\value{
Logical vector indicating which states are dark / stubborn.
}
\description{
Dark-State Flags
}
\author{
OpenAI Codex
}
