% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_levels_vec.R
\name{get_levels_vec}
\alias{get_levels_vec}
\title{get_levels_vec}
\usage{
get_levels_vec(num_opinions)
}
\arguments{
\item{num_opinions}{number of opinions}
}
\value{
levels
}
\description{
get_levels_vec
}
\examples{
#---------------------------------
# get_levels_vec (num_opinions)
#---------------------------------
}
\author{
Victor Bernal \email{victor.arturo.bernal@gmail.com}
}
