% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_helpers.R
\name{get_state_camp_index}
\alias{get_state_camp_index}
\title{State Camp Index}
\usage{
get_state_camp_index(num_opinions)
}
\arguments{
\item{num_opinions}{Number of opinion states.}
}
\value{
Integer vector where 1 = red camp and 2 = blue camp.
}
\description{
State Camp Index
}
\author{
OpenAI Codex
}
