% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_helpers.R
\name{get_state_labels}
\alias{get_state_labels}
\title{State Labels}
\usage{
get_state_labels(num_opinions)
}
\arguments{
\item{num_opinions}{Number of opinion states.}
}
\value{
Character vector of labels aligned with \code{get_levels_vec()}.
}
\description{
State Labels
}
\author{
OpenAI Codex
}
