% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visual_heatmap.R
\docType{package}
\name{heatmapPlot}
\alias{heatmapPlot}
\title{heatmapPlot}
\usage{
heatmapPlot(opinion_history, num_opinions)
}
\arguments{
\item{opinion_history}{opinion_history}

\item{num_opinions}{num_opinions}
}
\examples{
#---------------------------------
# heatmapPlot
# heatmapPlot(opinion_history, num_opinions)
#---------------------------------
}
\author{
Victor Bernal \email{victor.arturo.bernal@gmail.com}
}
