% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialization.R
\name{init_network}
\alias{init_network}
\title{Initialize Network}
\usage{
init_network(params)
}
\arguments{
\item{params}{Simulation parameters.}

\item{opinions}{Agent opinions.}

\item{B0}{Initial network adjacency matrix.}

\item{state}{Current simulation state.}

\item{trackers}{Simulation trackers.}

\item{network_state}{Network state.}

\item{stop_reason}{Simulation stop reason.}

\item{t}{Current simulation time.}

\item{comp_time}{Computed event time.}

\item{event_counter}{Event counter.}
}
\value{
Initialized network state.
}
\description{
Initialize the simulation network.
}
