% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackers.R
\name{init_trackers}
\alias{init_trackers}
\title{Track Simulation State}
\usage{
init_trackers(params, opinions)
}
\arguments{
\item{params}{Simulation parameters.}

\item{opinions}{Agent opinions.}

\item{t}{Current simulation time.}
}
\value{
Updated simulation trackers.
}
\description{
Update simulation trackers.
}
