% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_opinions.R
\docType{package}
\name{initialize_opinions}
\alias{initialize_opinions}
\title{Initialize voters' opinions}
\usage{
initialize_opinions(n)
}
\arguments{
\item{n}{number of individuals vertices}
}
\value{
initialize_opinions \code{x}.
}
\examples{
#---------------------------------
# Initialize opinions
# initialize_opinions(n = 20)
#---------------------------------
}
\references{
\url{}
}
\seealso{
\code{\link{brocolors}}
}
\author{
Victor Bernal \email{victor.arturo.bernal@gmail.com}
}
\keyword{hplot}
