% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_transition_matrix.R
\docType{package}
\name{make_transition_matrix}
\alias{make_transition_matrix}
\title{make_transition_matrix}
\usage{
make_transition_matrix(N)
}
\arguments{
\item{N}{}
}
\value{
transition matrix
}
\examples{
#---------------------------------
# make_transition_matrix
#---------------------------------
}
\references{
\url{}
}
\seealso{
\code{\link{brocolors}}
}
\author{
Victor Bernal \email{victor.arturo.bernal@gmail.com}
}
\keyword{hplot}
