% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_helpers.R
\name{opinion_state_index}
\alias{opinion_state_index}
\title{Opinion State Index}
\usage{
opinion_state_index(opinions, num_opinions)
}
\arguments{
\item{opinions}{Opinion values.}

\item{num_opinions}{Number of opinion states.}
}
\value{
Integer vector of state indices.
}
\description{
Opinion State Index
}
\author{
OpenAI Codex
}
