% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_helpers.R
\name{reconstruct_bipartite}
\alias{reconstruct_bipartite}
\title{Reconstruct a Bipartite Matrix From Membership Lists}
\usage{
reconstruct_bipartite(members, n, m)
}
\arguments{
\item{members}{List of group memberships.}

\item{n}{Number of individuals.}

\item{m}{Number of groups.}
}
\value{
Sparse bipartite matrix.
}
\description{
Reconstruct a Bipartite Matrix From Membership Lists
}
\author{
OpenAI Codex
}
