% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_helpers.R
\name{remove_value_fast}
\alias{remove_value_fast}
\title{Remove a Value From a Vector}
\usage{
remove_value_fast(vec, value)
}
\arguments{
\item{vec}{Integer vector.}

\item{value}{Value to remove.}
}
\value{
Input vector with the first match removed.
}
\description{
Remove a Value From a Vector
}
\author{
OpenAI Codex
}
