% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_helpers.R
\name{sample_outsider}
\alias{sample_outsider}
\title{Sample an Outsider For a Group}
\usage{
sample_outsider(group_i, members, groups_of_individual, n, max_tries = 25L)
}
\arguments{
\item{group_i}{Group index.}

\item{members}{Membership list by group.}

\item{groups_of_individual}{Reverse membership index.}

\item{n}{Number of individuals.}

\item{max_tries}{Number of rejection-sampling attempts.}
}
\value{
Integer individual id or \code{NA}.
}
\description{
Sample an Outsider For a Group
}
\author{
OpenAI Codex
}
