% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Schelling_step.R
\docType{package}
\name{schelling_step}
\alias{-package}
\alias{schelling_step}
\title{Single Schelling step}
\usage{
schelling_step(RIG, opinions, threshold)
}
\arguments{
\item{RIG}{number of individuals vertices}

\item{opinions}{number of group vertices}

\item{threshold}{threshold}
}
\value{
rig
}
\examples{
#---------------------------------
# A single Schelling step
# schelling_step(RIG, opinions, threshold)
#---------------------------------
}
\author{
Victor Bernal \email{victor.arturo.bernal@gmail.com}
}
