% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Schelling_step_bipartite.R
\docType{package}
\name{schelling_step_bipartite}
\alias{schelling_step_bipartite}
\title{Single Schelling step bipartite}
\usage{
schelling_step_bipartite(B, opinions, threshold)
}
\arguments{
\item{B}{bipartite grpah}

\item{opinions}{number of group vertices}

\item{threshold}{threshold}
}
\value{
B
}
\examples{
#---------------------------------
# A single Schelling step
# schelling_step(B, opinions, threshold)
#---------------------------------
}
\author{
Victor Bernal \email{victor.arturo.bernal@gmail.com}
}
