% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schelling_step_bipartite_gillespie.R
\docType{package}
\name{schelling_step_bipartite_gillespie_multi}
\alias{schelling_step_bipartite_gillespie_multi}
\title{Single Schelling step on the bipartite graph with gillespie}
\usage{
schelling_step_bipartite_gillespie_multi(
  bipartite_matrix,
  RIG,
  opinions,
  const = 1
)
}
\arguments{
\item{bipartite_matrix}{bipartite grpah}

\item{RIG}{RIG}

\item{opinions}{number of group vertices}

\item{const}{cont}
}
\value{
bipartite_matrix, RIG, dt
}
\examples{
#---------------------------------
# A single Schelling step
#schelling_step_bipartite_gillespie_multi (bipartite_matrix, RIG, opinions, const = 1)
#---------------------------------
}
\author{
Victor Bernal \email{victor.arturo.bernal@gmail.com}
}
