% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schelling_step_bipartite_random.R
\docType{package}
\name{schelling_step_bipartite_random}
\alias{schelling_step_bipartite_random}
\title{Single Schelling step_ on the bipartite graph with vertex at random}
\usage{
schelling_step_bipartite_random(
  bipartite_matrix,
  RIG,
  opinions,
  vertex,
  const = 1
)
}
\arguments{
\item{RIG}{RIG}

\item{opinions}{number of group vertices}

\item{vertex}{vertex index}

\item{const}{cont}

\item{B}{bipartite grpah}

\item{threshold}{threshold}
}
\value{
bipartite_matrix, RIG
}
\examples{
#---------------------------------
# A single Schelling step
# schelling_step_bipartite_random(bipartite_matrix, opinions, vertex)
#---------------------------------
}
\author{
Victor Bernal \email{victor.arturo.bernal@gmail.com}
}
