% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_voter.R
\docType{package}
\name{simulate_voter}
\alias{simulate_voter}
\title{Simulation voter's dynamics}
\usage{
simulate_voter(rig, opinions, times, kappa)
}
\arguments{
\item{rig}{number of individuals vertices}

\item{opinions}{number of group vertices}

\item{times}{number of iterations}

\item{kappa}{Poisson clock ringing rate (see \code{\link[graphics]{par}})}

\item{frac_pos}{fraction of positive opinions +1}
}
\value{
opinions \code{opinions}.
}
\examples{
#---------------------------------
# Simulate Voter
# simulate_voter(rig, opinions, times, kappa)
#---------------------------------
}
\author{
Victor Bernal \email{victor.arturo.bernal@gmail.com}
}
