% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visual_step_bipartite.R
\docType{package}
\name{visual_bipartite}
\alias{visual_bipartite}
\title{visual_bipartite}
\usage{
visual_bipartite(bipartite, opinions, num_opinions)
}
\arguments{
\item{bipartite}{biparitite}

\item{opinions}{number of group vertices}

\item{num_opinions}{number opinions}
}
\examples{
#---------------------------------
# visual_bipartite
# visual_bipartite(bipartite, opinions, num_opinions)
#---------------------------------
}
\author{
Victor Bernal \email{victor.arturo.bernal@gmail.com}
}
