% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visual_degree_time.R
\docType{package}
\name{visual_degree_time}
\alias{visual_degree_time}
\title{visual_degree_time}
\usage{
visual_degree_time(avg_deg)
}
\arguments{
\item{avg_deg}{avg_deg}
}
\examples{
#---------------------------------
# visual_degree_time
# visual_degree_time(avg_deg)
#---------------------------------
}
\author{
Victor Bernal \email{victor.arturo.bernal@gmail.com}
}
