% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visual_epidemic_camps.R
\name{visual_epidemic_camps}
\alias{visual_epidemic_camps}
\title{Plot Final Epidemic Outcomes By Camp}
\usage{
visual_epidemic_camps(final_camp_sir)
}
\arguments{
\item{final_camp_sir}{Matrix of final SIR counts by camp.}
}
\description{
Plot Final Epidemic Outcomes By Camp
}
\author{
OpenAI Codex
}
