% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visual_graph.R
\docType{package}
\name{visual_graph}
\alias{visual_graph}
\title{visual_graph}
\usage{
visual_graph(RIG, num_opinions)
}
\arguments{
\item{RIG}{RIG}

\item{num_opinions}{num_opinions}
}
\examples{
#---------------------------------
# visual_graph
# visual_graph(RIG, num_opinions)
#---------------------------------
}
\author{
Victor Bernal \email{victor.arturo.bernal@gmail.com}
}
