% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visual_histo_pergroup.R
\docType{package}
\name{visual_histo_pergroup}
\alias{visual_histo_pergroup}
\title{visual_histo_pergroup}
\usage{
visual_histo_pergroup(opinions, num_opinions, bipartite)
}
\arguments{
\item{opinions}{opinions}

\item{num_opinions}{number of opinions}

\item{bipartite}{bipartite}
}
\value{
visual_histo_pergroup \code{x}.
}
\examples{
#---------------------------------
# visual_histo_pergroup
# visual_histo_pergroup_multi(opinions, num_opinions, bipartite)
#---------------------------------
}
\references{
\url{}
}
\seealso{
\code{\link{brocolors}}
}
\author{
Victor Bernal \email{victor.arturo.bernal@gmail.com}
}
\keyword{hplot}
