% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visual_opinion_shares.R
\name{visual_opinion_shares}
\alias{visual_opinion_shares}
\title{Plot Final Opinion Shares}
\usage{
visual_opinion_shares(frac_mat, num_opinions)
}
\arguments{
\item{frac_mat}{Matrix of opinion fractions over time.}

\item{num_opinions}{Number of opinion states.}
}
\description{
Plot Final Opinion Shares
}
\author{
OpenAI Codex
}
