% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visual_sir_camp_time.R
\name{visual_sir_camp_time}
\alias{visual_sir_camp_time}
\title{Plot SIR Dynamics By Camp}
\usage{
visual_sir_camp_time(camp_sir_history, time_points = NULL)
}
\arguments{
\item{camp_sir_history}{List of camp-by-compartment SIR count matrices.}

\item{time_points}{Optional vector of recorded simulation times.}
}
\description{
Plot SIR Dynamics By Camp
}
\author{
OpenAI Codex
}
