% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visual_step.R
\docType{package}
\name{visual_step}
\alias{visual_step}
\title{visual_step}
\usage{
visual_step(RIG, opinions)
}
\arguments{
\item{RIG}{RIG}

\item{opinions}{number of group vertices}
}
\examples{
#---------------------------------
# visual_step
# visual_step(rig_graph, opinions )
#---------------------------------
}
\author{
Victor Bernal \email{victor.arturo.bernal@gmail.com}
}
