% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visual_step_time_SIR.R
\name{visual_step_time_SIR}
\alias{visual_step_time_SIR}
\title{visual_step_time_SIR}
\usage{
visual_step_time_SIR(x)
}
\arguments{
\item{x}{SIR data frame}
}
\description{
visual_step_time_SIR
}
\examples{
#---------------------------------
# visual_step_time_SIR(x)
#---------------------------------
}
\author{
Victor Bernal
}
