% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voter_step.R
\docType{package}
\name{voter_step}
\alias{voter_step}
\title{Single voter step}
\usage{
voter_step(rig, opinions, kappa, vertex)
}
\arguments{
\item{rig}{number of individuals vertices}

\item{opinions}{number of group vertices}

\item{kappa}{Poisson clock ringing rate (see \code{\link[graphics]{par}})}

\item{vertex}{vertex index}
}
\value{
opinions \code{opinions}.
}
\examples{
#---------------------------------
# A single voter step
# voter_step(rig, opinions, kappa)
#---------------------------------
}
\author{
Victor Bernal \email{victor.arturo.bernal@gmail.com}
}
