% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voter_step_gillespie.R
\docType{package}
\name{voter_step_gillespie}
\alias{voter_step_gillespie}
\title{Single voter_step_gillespie}
\usage{
voter_step_gillespie(rig, opinions, kappa)
}
\arguments{
\item{rig}{number of individuals vertices}

\item{opinions}{number of group vertices}

\item{kappa}{Poisson clock ringing rate (see \code{\link[graphics]{par}})}
}
\value{
opinions \code{opinions}.
}
\examples{
#---------------------------------
# voter_step_gillespie
# voter_step_gillespie(rig, opinions, kappa)
#---------------------------------
}
\author{
Victor Bernal \email{victor.arturo.bernal@gmail.com}
}
