% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voter_step_gillespie_multi.R
\docType{package}
\name{voter_step_gillespie_multi}
\alias{voter_step_gillespie_multi}
\title{Single voter_step_gillespie_multi}
\usage{
voter_step_gillespie_multi(
  rig,
  opinions,
  vertex = NULL,
  kappa,
  decision_matrix = NULL
)
}
\arguments{
\item{rig}{number of individuals vertices}

\item{opinions}{number of group vertices}

\item{vertex}{vertex}

\item{kappa}{Poisson clock ringing rate (see \code{\link[graphics]{par}})}

\item{decision_matrix}{decision_matrix}
}
\value{
opinions, dt \code{opinions}.
}
\examples{
#---------------------------------
# voter_step_gillespie_multi
# voter_step_gillespie_multi (rig, opinions, kappa)
#---------------------------------
}
\author{
Victor Bernal \email{victor.arturo.bernal@gmail.com}
}
