% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epidemic_moves.R
\name{apply_epidemic_event}
\alias{apply_epidemic_event}
\title{epidemic_event}
\usage{
apply_epidemic_event(state, rates, params, t)
}
\arguments{
\item{state}{Current epidemic state.}

\item{params}{Simulation parameters.}

\item{t}{Current simulation time.}

\item{opinions}{Current agent opinions.}

\item{comp_time}{Computed event time.}
}
\value{
Updated simulation state.
}
\description{
Process an epidemic event in the simulation.
}
