% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epidemic_moves.R
\name{apply_epidemic_event}
\alias{apply_epidemic_event}
\title{epidemic_event}
\usage{
apply_epidemic_event(state, rates, params, t)
}
\arguments{
\item{params}{Simulation parameters.}
}
\description{
epidemic_event
}
