% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bipartite_to_rig.R,
%   R/debug_illegal_transition.R
\docType{package}
\name{bipartite_to_rig}
\alias{bipartite_to_rig}
\alias{debug_illegal_transition}
\title{Bipartite to RIG}
\usage{
bipartite_to_rig(bipartite_matrix)

debug_illegal_transition(old_opinions, new_opinions, t = NULL)
}
\arguments{
\item{bipartite_matrix}{bipartite_matrix}

\item{old_opinions}{old_opinions}

\item{new_opinions}{new_opinions}

\item{t}{time}
}
\value{
Random Intersection Graph \code{rig}.
}
\description{
Bipartite to RIG

Detect illegal Voter moves
}
\examples{
#---------------------------------
# bipartite_to_rig(bipartite_matrix)
#---------------------------------
#---------------------------------
# debug_illegal_transition
#---------------------------------
}
\references{
\url{}

\url{}
}
\seealso{
\code{\link{brocolors}}

\code{\link{brocolors}}
}
\author{
Victor Bernal \email{victor.arturo.bernal@gmail.com}
}
\keyword{hplot}
