% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stopping.R
\name{check_stopping}
\alias{check_stopping}
\title{Stopping}
\usage{
check_stopping(state, params)
}
\arguments{
\item{state}{Current simulation state.}

\item{params}{Simulation parameters.}
}
\value{
Updated stopping state.
}
\description{
Determine whether the simulation should stop.
}
