% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stability.R
\name{create_stability_monitor}
\alias{create_stability_monitor}
\title{Stability monitoring}
\usage{
create_stability_monitor(
  window = stability_window,
  threshold = stability_threshold
)
}
\description{
Stability monitoring
}
