% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stability.R
\name{create_stability_monitor}
\alias{create_stability_monitor}
\title{Stability monitoring}
\usage{
create_stability_monitor(window = 100, threshold = 0.01, persistence = 5)
}
\description{
Stability monitoring
}
