% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_bipartite.R
\docType{package}
\name{generate_bipartite}
\alias{generate_bipartite}
\title{Generate bipartite matrix}
\usage{
generate_bipartite(n, m, individual_weights, group_weights)
}
\arguments{
\item{n}{number of individuals vertices}

\item{m}{number of group vertices}

\item{individual_weights}{ind vertex weight}

\item{group_weights}{(see \code{\link[graphics]{par}})}
}
\value{
bipartite matrix \code{x}.
}
\examples{
#---------------------------------
# Generate a bipartite
# generate_bipartite(n = 20, m = 4, individual_weights, group_weights, lambda)
#---------------------------------
}
\references{
\url{}
}
\seealso{
\code{\link{brocolors}}
}
\author{
Victor Bernal \email{victor.arturo.bernal@gmail.com}
}
\keyword{hplot}
