% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_opinions_multi.R
\docType{package}
\name{initialize_opinions_multi}
\alias{initialize_opinions_multi}
\title{Initialize voters' opinions multi}
\usage{
initialize_opinions_multi(n, num_opinions = NULL)
}
\arguments{
\item{n}{number of individuals vertices}

\item{num_opinions}{number of opinions}
}
\value{
initialize_opinions \code{x}.
}
\examples{
#---------------------------------
# Initialize opinions
# initialize_opinions_multi(n = 20,num_Opinions =2)
#---------------------------------
}
\references{
\url{}
}
\seealso{
\code{\link{brocolors}}
}
\author{
Victor Bernal \email{victor.arturo.bernal@gmail.com}
}
\keyword{hplot}
