% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconstruct_bipartite.R
\name{reconstruct_bipartite}
\alias{reconstruct_bipartite}
\title{Reconstruct a Bipartite Matrix from members}
\usage{
reconstruct_bipartite(members, n, m)
}
\arguments{
\item{members}{membership}

\item{n}{number of individuals}

\item{m}{number of groups}
}
\value{
Bipartite Matrix.
}
\description{
Reconstruct a Bipartite Matrix from members
}
\examples{
#---------------------------------
# reconstruct_bipartite(members, n, m)
#---------------------------------
}
\author{
Victor Bernal
}
