% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/social_moves.R
\name{sample_social_event}
\alias{sample_social_event}
\title{sample_social_event}
\usage{
sample_social_event(state, rates, params)
}
\arguments{
\item{params}{Simulation parameters.}
}
\description{
sample_social_event
}
