% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visual_histo_pergroup.R
\name{visual_histo_pergroup}
\alias{visual_histo_pergroup}
\title{visual_histo_pergroup}
\usage{
visual_histo_pergroup(opinions, num_opinions, members)
}
\arguments{
\item{opinions}{opinions}

\item{num_opinions}{number of opinions}

\item{members}{members}
}
\value{
visual_histo_pergroup \code{x}.
}
\description{
visual_histo_pergroup
}
\examples{
#---------------------------------
# visual_histo_pergroup
# visual_histo_pergroup_multi(opinions, num_opinions, members)
#---------------------------------
}
\author{
Victor Bernal \email{victor.arturo.bernal@gmail.com}
}
