% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visual_SIR.R
\name{visual_SIR}
\alias{visual_SIR}
\title{visual_SIR}
\usage{
visual_SIR(inf_time, inf_camp)
}
\arguments{
\item{inf_time}{Infection times.}

\item{inf_camp}{Infection camps.}
}
\description{
visual_SIR
}
\examples{
#---------------------------------
# visual_SIR(inf_time, inf_camp)
#---------------------------------
}
\author{
Victor Bernal
}
