% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visual_step_multi.R
\name{visual_step_multi}
\alias{visual_step_multi}
\title{visual_step_multi}
\usage{
visual_step_multi(RIG, opinions, num_opinions)
}
\arguments{
\item{RIG}{RIG}

\item{opinions}{number of group vertices}

\item{num_opinions}{number opinions}
}
\description{
visual_step_multi
}
\examples{
#---------------------------------
# visual_step_multi
# visual_step_multi(RIG, opinions, num_opinions)
#---------------------------------
}
\author{
Victor Bernal \email{victor.arturo.bernal@gmail.com}
}
