% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visual_step_time.R
\docType{package}
\name{visual_step_time}
\alias{visual_step_time}
\title{visual_step_time}
\usage{
visual_step_time(opinion_history, num_opinions)
}
\arguments{
\item{opinion_history}{opinion_history}

\item{num_opinions}{num_opinions}
}
\examples{
#---------------------------------
# visual_step_time
# visual_step_time(opinion_history, num_opinions)
#---------------------------------
}
\author{
Victor Bernal \email{victor.arturo.bernal@gmail.com}
}
