% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visual_step_time.R
\name{visual_step_time}
\alias{visual_step_time}
\title{visual_step_time}
\usage{
visual_step_time(frac_mat, num_opinions, n)
}
\arguments{
\item{frac_mat}{fraction of opinions}

\item{num_opinions}{number of opinions}

\item{n}{number of individuals}
}
\description{
visual_step_time
}
\examples{
#---------------------------------
# visual_step_time
# visual_step_time(frac_mat, num_opinions)
#---------------------------------
}
\author{
Victor Bernal \email{victor.arturo.bernal@gmail.com}
}
